name:邹稷 title：修改了库存统计的删除 date：2020/9/23

+++ src/main/java/com/yidu/dayDispose/service/impl/InventoryServiceImpl.java
@@ -320,7 +320,7 @@ public class InventoryServiceImpl implements InventoryService {
                         //期初标志 1.是   0.否
                         cashClosedPayInventory.setInitialSigns(0);
                         //总金额
-                        cashClosedPayInventory.setTotalMoney((int) caYSYFInventoryMapper.getToca());
+                        cashClosedPayInventory.setTotalMoney(caYSYFInventoryMapper.getToca());
 
                         cashClosedPaylnventoryMapper.insertCashClosedPaylnventory(cashClosedPayInventory);
                     }
